+++ repostbot/whitelist_status.py
@@ -5,3 +5,4 @@ class WhitelistAddStatus(Enum):
     SUCCESS = 1
     FAIL = 2
     ALREADY_EXISTS = 3
+    ADDED_AND_REMOVED = 4